[refactor] UIとロジックの分離
