+++ user/src/components/ApplicationForms/Stage/StageForm/StageForm.stories.tsx
@@ -1,0 +1,24 @@
+import { Meta, StoryObj } from '@storybook/react';
+import "@globals";
+import StageForm from './StageForm';
+
+export default {
+title: 'Components/StageForm',
+tags: ["autodocs"],
+component: StageForm,
+parameters: {
+    docs: {
+      source: {
+        type: "auto",
+      },
+    },
+  },
+} as Meta<typeof StageForm>;
+
+type Story = StoryObj<typeof StageForm>;
+
+export const Default: Story = {
+  args: {
+
+  },
+};
