+++ user/src/components/ApplicationForms/Stage/StageForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./StageForm";
